Refactored and tests pass
